Simplify discord bot
