--- discord-bot/handlers/commandHandler.js
@@ -1,20 +1,0 @@
-const { MessageFlags } = require('discord.js');
-
-module.exports = async (interaction, client) => {
-    const command = client.commands.get(interaction.commandName);
-    if (!command) {
-        console.error(`No command matching ${interaction.commandName} was found.`);
-        return;
-    }
-    try {
-        await command.execute(interaction);
-    } catch (error) {
-        console.error(`Error executing ${interaction.commandName}`, error);
-        const replyOptions = { content: 'There was an error executing this command!', flags: [MessageFlags.Ephemeral] };
-        if (interaction.replied || interaction.deferred) {
-            await interaction.followUp(replyOptions);
-        } else {
-            await interaction.reply(replyOptions);
-        }
-    }
-};
